utils/news2hugo/requirements.txt
markdown2==2.4.8
textile==4.0.2
ruamel.yaml==0.17.21
